chore(replay): move first-party-hosts into ml-mirror, its only consumer
Generated-By: PostHog Code
Task-Id: 273554de-31f0-49a3-b3d6-f94d9c8fa669
